+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## v1.4.5 (not released yet)
 - Remove esri-leaflet from the dependencies (but it's still needed)
+- Make the map view only update the title, if editable
 
 ## v1.4.4 - 05/04/2019
 - Rename a CSS class called "row" to avoid collisions with styles from the host app